move SponsorGroup to a separate component file
